requirements.txt
amaascore
business_calendar
python-dateutil
yahoo-finance